pyproject.toml
[project]
name = "telegram-forwarding-bot"
version = "1.0.0"
description = "Advanced Telegram bot with dual-client forwarding system"
authors = ["Bot Developer <dev@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=20.7",
    "telethon>=1.29.3",
    "sqlalchemy>=2.0.23",
    "alembic>=1.13.1",
    "asyncpg>=0.29.0",
    "redis>=5.0.1",
    "aioredis>=2.0.1",
    "celery>=5.3.4",
    "structlog>=23.2.0",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.7",
    "pydantic>=2.5.2",
    "pydantic-settings>=2.1.0",
    "prometheus-client>=0.19.0",
    "sentry-sdk>=1.39.2",
    "uvloop>=0.19.0"
]
